refactor: layout 중첩제외를 위한 chatRoom페이지 변경
