+++ src/app/chatrooms/_api/getChatRooms.ts
@@ -1,10 +1,10 @@
 import { authCheck } from "@/utils/function/authCheck";
 import { ChatRoomsResponse } from "@/utils/types/chat/chatRooms";
 
-const getChatRooms = async (): Promise<ChatRoomsResponse | null> => {
+const getChatRooms = async (): Promise<ChatRoomsResponse> => {
   const isTokenValid = authCheck();
 
-  if (!isTokenValid) return null;
+  if (!isTokenValid) throw new Error("401");
 
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/chat-rooms`,
