+++ src/app/chatrooms/layout.tsx
@@ -1,0 +1,7 @@
+export default function ProductsDetail({
+  children
+}: {
+  children: React.ReactNode;
+}) {
+  return <main className="px-2">{children}</main>;
+}
